Move token enum to a different file since it'll get a lot bigger later
