+++ src/lexer/mod.rs
@@ -1,26 +1,10 @@
+mod token;
+
 use super::definitions::{FileStream};
 use std::io::{Read, BufRead};
 use regex::Regex;
 
-
-
-
-
-pub enum Token {
-    Namespace(u64),
-    Use(u64), // Will probably need support for PHP7 soon
-
-    ClassStart(u64),
-    ClassEnd(u64),
-
-    MethodStart(u64),
-    MethodEnd(u64),
-
-    FunctionStart(u64),
-    FunctionEnd(u64)
-}
-
-
+pub use token::Token;
 
 
 pub struct Lexer {}
